backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.1.0
psycopg[binary,pool]==3.1.19
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.23.8
PyJWT[crypto]==2.9.0
authlib==1.4.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
